refactor TestCannotUnjailUnlessJailed to use simapp
